% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icph.R
\name{icph}
\alias{icph}
\title{Invariant Causal Prediction in the presence of latent variables}
\usage{
icph(D, E, target, number.of.states = 2, model = "HMM",
  intercept = TRUE, alpha = 0.05, par.icp = list(silent = FALSE,
  stopIfEmpty = TRUE, output.pvalues = TRUE), par.test = list(silent =
  TRUE))
}
\arguments{
\item{D}{A numeric \code{n} times \code{d} data matrix. Each column corresponds to a variable, each row to an observation.}

\item{E}{The environment indicator, can be numeric or a factor variable. 
This variable indicates which environment or interventional setting the observations belong to, i.e., 
the \code{i}'th entry of \code{E} should be \code{k} if the observation in the 
\code{i}'th row of the data matrix \code{D} comes from environment \code{k}.}

\item{target}{Integer between 1 and d indicating which of the variables 
in the data matrix \code{D} should be taken as target variable.}

\item{number.of.states}{An integer greater or equal to 2 specifying the number of mixture components. 
Numerical experiments show poor performance of the test for more than 3 components.}

\item{model}{The dependence model for the hidden variables. If ``IID'', 
all observations are assumed to be independent. 
If ``HMM'', the hidden variables are assumed to follow a first order Markov chain.}

\item{intercept}{Logical indicating whether an intercept should 
be included in the linear regression of all components.}

\item{alpha}{Numeric between 0 and 1 indicating the desired type I error control.}

\item{par.icp}{A list containing the following options. 
\itemize{
  \item{\code{max.parents}:}{ Integer between 1 and \code{d-1}, indicating the maximal size 
    of parent sets that should be tested for.}
  \item{\code{silent}:}{ Logical. If \code{TRUE}, then information on the progression of the algorithm 
is printed to the console.}
  \item{\code{stopIfEmpty}:}{ Logical. If \code{FALSE}, then the algorithm terminates if the 
empty set is accepted.}
  \item{\code{output.pvalues}:}{ Logical. If \code{TRUE}, p-values for the individual hypotheses 
will be calculated and returned. Since p-values are computed by repeated function evaluations 
along a grid of thresholds, this option is time consuming if the data matrix \code{D} is large. 
If \code{FALSE}, the individual hypotheses are only tested at level \code{alpha} 
(without calculating an actual p-value)}
}}

\item{par.test}{A list of options to be passed to the testing procedure. 
See documentation of \link{test.equality.sr} for content and default values.}
}
\value{
Object of class \code{icph} containg a list with the following items. 
    \item{parent.set}{ A subset of the indices from 1 to \code{d} (except \code{target}) 
      indicating the estimated set of observable causal parents.}
    \item{pvalues}{ A data frame with the results for the individual hypothesis tests. 
      If \code{output.pvalues = TRUE}, the p-values are returned. 
      Otherwise, an indicator for the rejection/accept of the individual hypothesis is returned.
    }
}
\description{
Infers the set of causal parents of a target variable Y among a vector of covariates X. The method allows for the presence of discrete latent variables acting on Y.
}
\examples{
 
  n <- 600
  # Binary hidden variable that follows an order 1 Markov chain
  Gamma <- matrix(c(.9, .1, .1, .9),2,2, byrow=TRUE)
  H <- rep(1, n)
  for(t in 2:n){
    H[t] <- sample(1:2, size=1, prob = Gamma[H[t-1],])
  }
  # SCM with assignments that change between every observation
  # True set of observable parents: S*={3}
  X4 <- H * rnorm(n)
  X3 <- H + cos((1:n)*2*pi/n)*X4 + .5*rnorm(n)
  Y <- (H==1)*(1 + X3 + .5*rnorm(n)) + (H==2)*(1 + 2*X3 + .7*rnorm(n))
  X2 <- .5 - X3 + exp(-3*(1:n)/n)*Y + .7*rnorm(n)
  # Data matrix
  D <- cbind(Y, X2, X3, X4)
  # Construct environments from time-ordering of data
  E <- rep(1:3, each=n/3)
  # Run icph
  res.icph <- icph(D=D, E=E, target=1)
  res.icph
  
}
\seealso{
\code{\link{test.equality.sr}}
}
\author{
Rune Christiansen \email{krunechristiansen@math.ku.dk}
}
